Updated the CMAKE File to not be redundant with the include DIRs, and Moved the time.h and time.cpp files back to where they belong!
