+++ CMakeLists.txt
@@ -79,8 +79,6 @@ INCLUDE(${3DEngineCpp_CMAKE_DIR}/FindASSIMP.cmake)
 
 # Define the include DIRs
 include_directories(
-	${3DEngineCpp_SOURCE_DIR}/headers
-	${3DEngineCpp_SOURCE_DIR}/sources
 	${OPENGL_INCLUDE_DIRS}
 	${GLEW_INCLUDE_DIRS}
 	${SDL2_INCLUDE_DIRS}
